Extracted QUERIES into separate file
